refactor: break common utils into separate files
